requirements.txt
ccxt>=4.2.0
pandas>=1.5.3
numpy>=1.23.0
pyyaml>=6.0
requests>=2.31.0